Move canvas specific global rendering code to canvas-layer.js
This introduces a new object, canvas-layer.js, responsible for driving canvas rendering.  It collects canvas specific methods from DrawManager and rectManager.

It also renames many methods to indicate that they're private.
